Start documenting LZ78 SPA
